ex03/src/main/java/org/zerock/mapper/BoardMapper.java
package org.zerock.mapper;

import java.util.List;

import org.apache.ibatis.annotations.Param;
import org.apache.ibatis.annotations.Select;
import org.zerock.domain.BoardVO;
import org.zerock.domain.Criteria;

public interface BoardMapper {
	public List<BoardVO> getList(Criteria criteria);
	
	
	public void insert(BoardVO boardVO);
	
	public void insertSelectKey(BoardVO boardVO);
	
	public BoardVO read(Long bno);
	
	public int delete(Long bno);
	
	public int update(BoardVO boardVO);
	
	public int getTotalCount(Criteria criteria);
	
	public List<BoardVO> getListWithPaging(Criteria criteria);
	
	public void updateReplyCnt(@Param("bno") Long bno, @Param("amount") int amount);
}
